+++ CHANGELOG.md
@@ -11,6 +11,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 // TODO: move to petgraph-algo
 - `dag_to_toposorted_adjacency_list` now requires `SafeCast<usize>`
+- Renamed feature `matrix_graph` to `matrix-graph`
+- Renamed feature `stable_graph` to `stable-graph`
+- `CSR` is now gated behind the `csr` feature
+- `AdjacencyMatrix` is now gated behind the `adjacency-matrix` feature
 
 ### Changed